+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Voice: Websocket Connections can now include custom authorization headers
 - Voice: Answer Webhook adds support for SIP User-To-User incoming headers
 - Voice: Add support for `shaken` signing on calls
+- Voice: Added support for `sip_code` on incoming event webhooks
 - Verify: Deprecates the `sandbox` parameter
 - Number Insights: Internally switched to supporting only Basic Authorization header authentication
 - Accounts: Internally switched to supporting only Basic Authorization header authentication